Move OauthApplicationRegistration to oauth_views.py
