+++ corehq/apps/hqwebapp/oauth_views.py
@@ -1,14 +1,82 @@
+from django.forms import modelform_factory
+from django.http import HttpResponseRedirect
 from django.utils.decorators import method_decorator
-
 from memoized import memoized
 from oauth2_provider.models import get_application_model
 from oauth2_provider.scopes import get_scopes_backend
 from oauth2_provider.views.base import AuthorizationView
 
+from corehq.apps.domain.decorators import (
+    require_superuser,
+)
 from corehq.apps.domain.models import Domain
 from corehq.apps.hqwebapp.decorators import use_bootstrap5
-from corehq.apps.hqwebapp.forms import HQAllowForm
+from corehq.apps.hqwebapp.forms import (
+    HQAllowForm,
+)
+from corehq.apps.hqwebapp.models import HQOauthApplication
 from corehq.apps.hqwebapp.oauth_scopes import DOMAIN_SCOPE_PREFIX
+from corehq.apps.hqwebapp.views import BasePageView
+from corehq.util.view_utils import reverse
+
+
+@method_decorator(require_superuser, name="dispatch")
+class OauthApplicationRegistration(BasePageView):
+    urlname = 'oauth_application_registration'
+    page_title = "Oauth Application Registration"
+    template_name = "hqwebapp/bootstrap3/oauth_application_registration_form.html"
+
+    @property
+    def page_url(self):
+        return reverse(self.urlname)
+
+    @property
+    def base_application_form(self):
+        return modelform_factory(
+            get_application_model(),
+            fields=(
+                "name",
+                "client_id",
+                "client_secret",
+                "client_type",
+                "authorization_grant_type",
+                "redirect_uris",
+            ),
+        )
+
+    @property
+    def hq_application_form(self):
+        return modelform_factory(
+            HQOauthApplication,
+            fields=(
+                "pkce_required",
+            ),
+        )
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        if 'forms' not in kwargs:
+            context['forms'] = [self.base_application_form(), self.hq_application_form()]
+        return context
+
+    def post(self, request, *args, **kwargs):
+
+        base_application_form = self.base_application_form(data=self.request.POST)
+        hq_application_form = self.hq_application_form(data=self.request.POST)
+
+        if base_application_form.is_valid() and hq_application_form.is_valid():
+            base_application_form.instance.user = self.request.user
+            base_application = base_application_form.save()
+            HQOauthApplication.objects.create(
+                application=base_application,
+                **hq_application_form.cleaned_data,
+            )
+        else:
+            return self.render_to_response(self.get_context_data(
+                forms=[base_application_form, hq_application_form]
+            ))
+
+        return HttpResponseRedirect(reverse('oauth2_provider:detail', args=[str(base_application.id)]))
 
 
 @method_decorator(use_bootstrap5, name="dispatch")
